Fixed classmap detection
Previously, any class that Composer passed to its `findFileWithExtension` method
was listed in the Composer classmap details. However, this was incorrect; later
autoloaders -- such as ZF's -- might be able to handle the given class.

Composer communicates a class miss with a null value; this patch detects that
null value and returns early in such situations, preventing invalid additions to
the classmap details.

Additionally, this has been cleaned up to follow PSR-1/PSR-2 and for internal
consistency (originally there was a mix of `array()` and `[]` usage; not it uses
solely `[]` -- since Z-Ray is only present on systems running PHP 5.4+ anyways).